requirements.txt
discord.py==2.4.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
pillow==11.0.0
python-dotenv==1.0.1
